+++ grails-app/controllers/bn/elicitator/output/JsonOutputGraph.groovy
@@ -9,7 +9,7 @@ class JsonOutputGraph extends OutputGraph {
 	Set<Variable> variables = []
 
 	@Override
-	void addEdge(Variable parent, Variable child, Float strength, Integer numUsers, Integer totalUsers) {
+	void addEdge(Variable parent, Variable child, double strength, int numUsers, int totalUsers) {
 		totalEdges ++
 		variables.add( parent )
 		variables.add( child )
